requirements.txt
# Core FastAPI dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Database
sqlalchemy==2.0.25

# HTTP and utilities
requests==2.31.0
python-dotenv==1.0.0

# Monitoring
prometheus-client==0.19.0

# Timezone
pytz==2025.2

# Financial data - using compatible version for Python 3.9
yfinance==0.2.36

# CLI and formatting
rich==13.7.0

# Data processing
numpy==1.26.0
pandas==2.1.0

# System monitoring
psutil==5.9.8

# RSS Feed parsing
feedparser==6.0.10
